+++ src/Resources/contao/languages/de/tl_fernschach_spieler_mails.php
@@ -1,0 +1,37 @@
+<?php 
+
+/**
+ * Contao Open Source CMS
+ * Copyright (C) 2005-2014 Leo Feyer
+ *
+ */
+
+/**
+ * Backend-Modul Übersetzungen
+ */
+
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['main']= array('E-Mail', 'E-Mail anlegen');
+
+// Standardfunktionen
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['new']= array('Neue E-Mail', 'Neue E-Mail anlegen');
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['edit']= array("E-Mail %s bearbeiten", "E-Mail %s bearbeiten");
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['copy']= array("E-Mail %s kopieren", "E-Mail %s kopieren");
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['delete']= array("E-Mail %s löschen", "E-Mail %s löschen");
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['send']= array("E-Mail %s versenden", "E-Mail %s versenden");
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['show']= array("Details zur E-Mail %s anzeigen", "Details zur E-Mail %s anzeigen");
+
+// Formularfelder
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['text_legend'] = 'Betreff und Text';
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['subject']= array('Betreff', 'Betreffzeile der E-Mail');
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['content']= array('Text', 'Textinhalt der E-Mail');
+
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['signatur_legend'] = 'Signatur';
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['signatur']= array('Signatur aktivieren', 'Signatur für diese E-Mail aktivieren.');
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['signatur_text']= array('Signaturtext', 'Die Signatur kann im Profil voreingestellt werden.');
+
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['mail_legend'] = 'E-Mail-Optionen';
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['copyBenutzer']= array('E-Mail-Kopie an mich', 'Eine E-Mail-Kopie wird an mich (aktueller Benutzer) gesendet.');
+
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['sent_state']= array('Versandstatus', 'Versandstatus');
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['sent_date']= array('Abgesendet am', 'Versanddatum');
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['sent_text']= array('Gesendeter Text', 'Gesendeter Text');